+++ src/views/CommitPanelViewProvider.ts
@@ -1686,7 +1686,10 @@ export class CommitPanelViewProvider implements vscode.WebviewViewProvider {
                 await runGitOperationFromPanel(this.actionDepsForRuntime(scopedRuntime()), "sync");
                 break;
             case "openRepository":
-                await vscode.commands.executeCommand("intelligit.openRepository");
+                await vscode.commands.executeCommand(
+                    "intelligit.openRepository",
+                    scopedRuntime()?.repository.root,
+                );
                 break;
             case "selectCommit":
                 this._onCommitSelected.fire(assertGitHash(msg.hash, "hash"));