+++ src/views/UndockedViewProvider.ts
@@ -1189,7 +1189,10 @@ export class UndockedViewProvider {
                 await runGitOperationFromPanel(actionDeps, msg.type);
                 break;
             case "openRepository":
-                await vscode.commands.executeCommand("intelligit.openRepository");
+                await vscode.commands.executeCommand(
+                    "intelligit.openRepository",
+                    this.selectedRepositoryRoot,
+                );
                 break;
             case "publishBranch":
                 await publishBranchFromPanel(fileActionDeps);